refactor(search): move fixed filters
